requirements.txt
colorama==0.4.4
configparser==5.2.0
psutil==5.8.0
python-telegram-bot==13.9
pyyaml==6.0
requests==2.26.0
python-i18n==0.3.9
matplotlib >= 3.4.2
python-binance >= 1.0.10
numpy >= 1.20.3
